lots of initial work changes.
I'm aware I'm working over a lot of things. Some of this is me wanting to do things my way, some of it is me wanting to make sure we have something that we could show at the end of the sprint. At the end of this, I intend to go back and try to incoporate what Karl was originally doing so that this would be better, but I want to make sure we hit the sprint goal first.