A lot more changes
Redid all the containers and visuals so that they extend the AbstractTile component, meaning that they no longer override the render function, but instead override the renderImpl of the AbstractTile
This also allows anything to be passed a card boolean which will wrap the visual/container in a card

Abstract Visual has a class in it for AxisConfig objects...
This thing is probably going to explode in complexity really quickly...

Bunch of changes to AreaGraph and LineGraph in order to accomidate passing a lot of values. Should really look into abstracting a lot of this stuff.